+++ bin/crashers/panic.go
@@ -1,0 +1,8 @@
+package main
+
+import "runtime/debug"
+
+func main() {
+	debug.SetTraceback("crash")
+	panic("panic")
+}