+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Use a QueryLink component to handle internal navigation via history
 
 ## [v0.9.0](https://github.com/elwinar/rcoredump/releases/tag/v0.9.0) - 2020-04-12
 ### Added